Rename MainActivity to UpcomingEventsActivity and move
- Rename MainActivity to UpcomingEventsActivity
- Add activity alias to the AndroidManifest
- Move the activity to the upcoming package
